:construction: Transferred functions from playback to sidebar component
