+++ src/app/components/sidebar/sidebar.component.html
@@ -4,7 +4,7 @@
       <i id="menu-button-icon-mob" class="material-icons-outlined menu-button-icon">close</i>
     </div>
     <div id="sidebar-leave" class="sidebar-header-leave">
-      <i class="material-icons-outlined sidebar-leave-button">logout</i>
+      <i class="material-icons-outlined sidebar-leave-button" (click)="leave()">logout</i>
     </div>
   </div>
   <div class="sidebar-content">
